Read splittable LZO effectively in transformer
Reading indexed LZO compressed files from Spark cannot be achieved using
the default `textFile()` because `textFile()` does not understand how to
split the LZO file even when an index is available.

This results in reading an indexed LZO file from the same executor, thus
reducing the job parallelism.

To read effectively indexed LZO, `newAPIHadoopFile()` is used to rely on
the available indexing, and splitting the file multiple partition
automatically.

The compression format of input files is passed as a CLI argument to
control which implementation is used.